refactor(stella-pipeline): move #1540's growth out of three god files, back under their pre-#1540 ceilings
Pure relocation, no behavior change — the extraction #1545 asked for,
taken to the natural module boundaries rather than the minimum line
count:

- candidate_ws.rs 1708 -> 1464: winner-only adoption (adopt_inner,
  attribute_conflict, blob_id, ADOPT_CACHE_EXCLUDES) moves into
  candidate_ws/adopt.rs beside the PathDivergence classifier and the
  parsers it feeds. The file drops below 1500 and leaves the
  grandfather baseline entirely; the AGENTS.md and crate-README
  god-file tables shed the row in the same change.
- pipeline.rs 3894 -> 3873: the uncorroborated-pass abstention stamp
  becomes verify::uncorroborated_pass_evidence beside its sibling
  evidence constructors, carrying the full rationale as rustdoc (and
  fixing a stale ModelJudge -> ModelVerdict rung name while moving).
- pipeline/tests.rs 2678 -> 2638: the headless scope-bypass scenario
  moves to tests/scope_gate_interactive.rs, which already owns both
  sides of the scope-review seam; its module doc now says so.

The baseline entries re-tighten to the new actuals (the gate
auto-accepts shrinkage), including drift other merges had left slack
in agent/tests.rs, command_deck.rs, driver.rs, deck_render.rs and the
bench harness.

Verification: relocation only — every moved test travels intact, and
cargo test -p stella-cli -p stella-pipeline plus clippy run post-push
on the branch; the file-size guard is the witness for the ceilings.

Closes #1545
