+++ crates/stella-pipeline/src/pipeline/tests/scope_gate_interactive.rs
@@ -1,8 +1,9 @@
-//! The interactive scope-review seam: a run that is NOT headless and carries a
-//! real approval gate must raise the card and, on approve, execute — the
-//! posture the command deck adopted once it grew a native approval surface.
-//! Its headless twin (`ScopeReviewRequiredHeadless`) lives in the parent
-//! module; this is the branch that had no coverage.
+//! The scope-review seam. Interactively: a run that is NOT headless and
+//! carries a real approval gate must raise the card and, on approve, execute —
+//! the posture the command deck adopted once it grew a native approval
+//! surface. Headlessly: a gate that cannot ask either refuses up front
+//! (naming `--plan`) or, under `headless_bypass_scope_review`, proceeds
+//! rather than consulting a gate that always aborts.
 
 use super::*;
 
@@ -507,5 +508,45 @@ async fn plan_mode_refuses_a_headless_run_rather_than_bypassing_the_gate() {
             AgentEvent::Error { message, .. } if message.contains("--plan")
         )),
         "the refusal must say why, naming the flag: {events:?}"
+    );
+}
+
+/// The headless approval port is `AlwaysAbortGate`, so "bypass scope review"
+/// has to mean *proceed*. Running the review anyway would hand the plan to a
+/// gate that always says no, empty it, and end the turn having done nothing —
+/// the same zero-work outcome as the hard error, just spelled differently.
+#[tokio::test]
+async fn headless_scope_bypass_proceeds_instead_of_asking_a_gate_that_always_aborts() {
+    // Six steps clears the default 5-step threshold, so scope review fires.
+    let provider = ScriptedProvider::new(vec![
+        text_result("CLASS: multi\nWITNESS: no\nVERIFIER: no"),
+        text_result(r#"["a","b","c","d","e","f"]"#),
+        text_result("done"),
+    ]);
+    let config = PipelineConfig {
+        headless: true,
+        headless_bypass_scope_review: true,
+        ..PipelineConfig::default()
+    };
+    let (outcome, events, _) = run_unisolated_with_router(
+        &provider,
+        config,
+        "Refactor across the codebase and then update all callers",
+        router(),
+    )
+    .await;
+    let outcome = outcome.expect("bypass must not surface the headless scope error");
+
+    assert_ne!(
+        outcome.status,
+        PipelineStatus::Aborted {
+            reason: "aborted at scope review".to_string()
+        },
+        "a bypassed review must not abort at the gate it bypassed"
+    );
+    let s = stages(&events);
+    assert!(
+        s.contains(&StageKind::Execute),
+        "the run reaches execute rather than ending at the gate: {s:?}"
     );
 }
