+++ scripts/file-size-baseline.txt
@@ -7,28 +7,27 @@
 # must be split instead. Raising an existing ceiling is allowed but is
 # never silent — it lands as a visible diff here, to be justified in
 # review like any other change.
-2302 bench/harbor_adapter/stella_harbor/__init__.py
+2231 bench/harbor_adapter/stella_harbor/__init__.py
 4292 bench/harbor_adapter/stella_harbor/secure_launcher.py
-2339 bench/harbor_adapter/tests/test_adapter.py
+2333 bench/harbor_adapter/tests/test_adapter.py
 3360 bench/harbor_adapter/tests/test_secure_launcher.py
 8272 bench/terminal_bench_analysis/tb21_analysis.py
 1911 bench/terminal_bench_analysis/tb21_evidence_contract.py
 4335 bench/terminal_bench_analysis/tests/test_tb21_analysis.py
 1659 bench/terminal_bench_analysis/tests/test_tb21_evidence_contract.py
 2270 crates/stella-cli/src/agent.rs
-1754 crates/stella-cli/src/agent/tests.rs
-1708 crates/stella-cli/src/candidate_ws.rs
-4757 crates/stella-cli/src/command_deck.rs
+1749 crates/stella-cli/src/agent/tests.rs
+4756 crates/stella-cli/src/command_deck.rs
 1506 crates/stella-cli/src/fleet_cmd.rs
 2129 crates/stella-core/src/bus.rs
-2770 crates/stella-core/src/driver.rs
+2747 crates/stella-core/src/driver.rs
 3675 crates/stella-core/src/driver/tests.rs
 1773 crates/stella-model/src/anthropic/tests.rs
 2076 crates/stella-model/src/openai.rs
 1542 crates/stella-model/src/zai.rs
 1815 crates/stella-model/src/zai/tests.rs
-3894 crates/stella-pipeline/src/pipeline.rs
-2678 crates/stella-pipeline/src/pipeline/tests.rs
+3873 crates/stella-pipeline/src/pipeline.rs
+2638 crates/stella-pipeline/src/pipeline/tests.rs
 2950 crates/stella-protocol/src/event.rs
 2074 crates/stella-store/src/lib.rs
 2268 crates/stella-store/src/tests.rs
@@ -37,7 +36,7 @@
 2308 crates/stella-tools/src/registry.rs
 1839 crates/stella-tools/src/scripts.rs
 1528 crates/stella-tui/src/deck.rs
-1581 crates/stella-tui/src/deck_render.rs
+1580 crates/stella-tui/src/deck_render.rs
 4028 crates/stella-tui/src/deck_ui.rs
 1665 crates/stella-tui/src/views/engine.rs
 1619 crates/stella-tui/src/views/session.rs
